preparando create/edit categoria

+++ StoreCollection/src/main/java/com/proyecto/StoreCollection/dto/request/CategoriaRequest.java
@@ -9,5 +9,5 @@ public class CategoriaRequest {
     private String nombre;
     @NotBlank
     private String slug;
-    private Long tiendaId;
+    private Integer tiendaId;
 }
